requirements.txt
gradio
transformers==4.29.2
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
datasets
pillow
pandas
tqdm
faiss-cpu
medmnist
